fixed structure of middlewares
