+++ middlewares/validateBody.js
@@ -1,0 +1,13 @@
+const { HttpError } = require("../helpers/index");
+const validateBody = (schema) => {
+  return (req, res, next) => {
+    const { error } = schema.validate(req.body);
+    if (error) {
+      next(new HttpError(400, error.message));
+    }
+
+    return next();
+  };
+};
+
+module.exports = validateBody;
