+++ models/user.js
@@ -20,7 +20,6 @@ const schema = new Schema(
       type: String,
       default: null,
     },
-    contacts: [{ type: Schema.Types.ObjectId, ref: "contact" }],
   },
   {
     versionKey: false,
