+++ server.js
@@ -1,6 +1,7 @@
+require("dotenv").config();
+
 const app = require("./app");
 const mongoose = require("mongoose");
-require("dotenv").config();
 
 mongoose.set("strictQuery", false);
