+++ services/users.js
@@ -4,8 +4,6 @@ const { HttpError } = require("../helpers/index");
 const bcrypt = require("bcrypt");
 const jwt = require("jsonwebtoken");
 
-require("dotenv").config();
-
 const { SECRET_KEY } = process.env;
 
 const createUser = async (userData) => {
